requirements.txt
flask
python-dotenv
openai
langchain
langchain-community
tiktoken
pypdf
pillow
pytesseract
